decimal type on ration

+++ src/users/entities/user.entity.ts
@@ -59,7 +59,7 @@ export class Users {
 
   @Column({
     nullable: false,
-    type: 'int',
+    type: 'decimal',
     width: 6,
   })
   ratio: number;
